+++ lib/nanoc/extra.rb
@@ -4,6 +4,7 @@ module Nanoc::Extra
 
   autoload 'AutoCompiler',      'nanoc/extra/auto_compiler'
   autoload 'CHiCk',             'nanoc/extra/chick'
+  autoload 'Pruner',            'nanoc/extra/pruner'
   autoload 'Validators',        'nanoc/extra/validators'
 
   # Deprecated; use {Nanoc::Context} instead